chore: api(service) module 분리
