+++ module-api/build.gradle
@@ -1,0 +1,15 @@
+
+bootJar {
+    enabled=false
+}
+
+jar {
+    enabled=true
+}
+
+dependencies {
+    implementation project(":module-domain")
+    implementation project(":module-common")
+
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+}
