+++ module-api/src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -1,0 +1,36 @@
+package kitchenpos.menugroup.application;
+
+import kitchenpos.menu.dto.MenuResponse;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.dto.MenuGroupRequest;
+import kitchenpos.menugroup.dto.MenuGroupResponse;
+import kitchenpos.menugroup.repository.MenuGroupRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+public class MenuGroupService {
+    private final MenuGroupRepository menuGroupRepository;
+
+    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+        this.menuGroupRepository = menuGroupRepository;
+    }
+
+    @Transactional
+    public MenuGroupResponse create(final MenuGroupRequest menuGroupRequest) {
+        return MenuGroupResponse.from(menuGroupRepository.save(menuGroupRequest.toEntity()));
+    }
+
+    @Transactional(readOnly=true)
+    public List<MenuGroupResponse> list() {
+
+        List<MenuGroup> menuGroups = menuGroupRepository.findAll();
+
+        return menuGroups.stream()
+                .map(MenuGroupResponse::from)
+                .collect(Collectors.toList());
+    }
+}
