+++ module-api/src/main/java/kitchenpos/menugroup/dto/MenuGroupResponse.java
@@ -1,0 +1,41 @@
+package kitchenpos.menugroup.dto;
+
+import kitchenpos.menugroup.domain.MenuGroup;
+
+import java.util.Objects;
+
+public class MenuGroupResponse {
+
+    private final Long id;
+    private final String name;
+
+    public MenuGroupResponse(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public static MenuGroupResponse from(MenuGroup menuGroup) {
+        return new MenuGroupResponse(menuGroup.getId(), menuGroup.getName());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        MenuGroupResponse that = (MenuGroupResponse) o;
+        return Objects.equals(id, that.id) && Objects.equals(name, that.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, name);
+    }
+}
