+++ module-api/src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,0 +1,39 @@
+package kitchenpos.product.application;
+
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.dto.ProductRequest;
+import kitchenpos.product.dto.ProductResponse;
+import kitchenpos.product.repository.ProductRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+public class ProductService {
+    private final ProductRepository productRepository;
+
+    public ProductService(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    @Transactional
+    public ProductResponse create(final ProductRequest productRequest) {
+        return ProductResponse.from(productRepository.save(new Product(productRequest.getName(), productRequest.getPrice())));
+    }
+    @Transactional(readOnly=true)
+    public List<ProductResponse> list() {
+
+        List<Product> products = productRepository.findAll();
+        return products.stream()
+                .map(ProductResponse::from)
+                .collect(Collectors.toList());
+    }
+
+    @Transactional(readOnly=true)
+    public Product getProduct(Long productId) {
+        return productRepository.findById(productId).orElseThrow(() -> new IllegalArgumentException("상품이 없습니다."));
+    }
+
+}
