+++ module-api/src/main/java/kitchenpos/product/dto/ProductResponse.java
@@ -1,0 +1,49 @@
+package kitchenpos.product.dto;
+
+import kitchenpos.product.domain.Product;
+
+import java.math.BigDecimal;
+import java.util.Objects;
+
+public class ProductResponse {
+
+    private final Long id;
+    private final String name;
+    private final BigDecimal price;
+
+    public ProductResponse(Long id, String name, BigDecimal price) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+    }
+
+    public static ProductResponse from(Product product) {
+        return new ProductResponse(product.getId(), product.getName(), product.getPrice());
+    }
+
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        ProductResponse that = (ProductResponse) o;
+        return Objects.equals(id, that.id) && Objects.equals(name, that.name) && Objects.equals(price, that.price);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, name, price);
+    }
+}
