+++ module-api/src/main/java/kitchenpos/table/application/TableGroupCreatedEventHandler.java
