--- src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -1,28 +1,0 @@
-package kitchenpos.menugroup.application;
-
-import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.dto.MenuGroupRequest;
-import kitchenpos.menugroup.repository.MenuGroupRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-public class MenuGroupService {
-    private final MenuGroupRepository menuGroupRepository;
-
-    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
-        this.menuGroupRepository = menuGroupRepository;
-    }
-
-    @Transactional
-    public MenuGroup create(final MenuGroupRequest menuGroupRequest) {
-        return menuGroupRepository.save(menuGroupRequest.toEntity());
-    }
-
-    @Transactional(readOnly=true)
-    public List<MenuGroup> list() {
-        return menuGroupRepository.findAll();
-    }
-}
