--- src/main/java/kitchenpos/menugroup/dto/MenuGroupResponse.java
@@ -1,26 +1,0 @@
-package kitchenpos.menugroup.dto;
-
-import kitchenpos.menugroup.domain.MenuGroup;
-
-public class MenuGroupResponse {
-
-    private final Long id;
-    private final String name;
-
-    public MenuGroupResponse(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public static MenuGroupResponse from(MenuGroup menuGroup) {
-        return new MenuGroupResponse(menuGroup.getId(), menuGroup.getName());
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
