--- src/main/java/kitchenpos/order/dto/OrderRequest.java
@@ -1,30 +1,0 @@
-package kitchenpos.order.dto;
-
-import kitchenpos.order.domain.OrderStatus;
-
-import java.util.List;
-
-public class OrderRequest {
-
-    private final Long orderTableId;
-    private final OrderStatus orderStatus;
-    private final List<OrderLineItemRequest> orderLineItemRequests;
-
-    public OrderRequest(Long orderTableId, OrderStatus orderStatus, List<OrderLineItemRequest> orderLineItemRequests) {
-        this.orderTableId = orderTableId;
-        this.orderStatus = orderStatus;
-        this.orderLineItemRequests = orderLineItemRequests;
-    }
-
-    public Long getOrderTableId() {
-        return orderTableId;
-    }
-
-    public List<OrderLineItemRequest> getOrderLineItemRequests() {
-        return orderLineItemRequests;
-    }
-
-    public OrderStatus getOrderStatus() {
-        return orderStatus;
-    }
-}
