--- src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,33 +1,0 @@
-package kitchenpos.product.application;
-
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.dto.ProductRequest;
-import kitchenpos.product.repository.ProductRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-public class ProductService {
-    private final ProductRepository productRepository;
-
-    public ProductService(ProductRepository productRepository) {
-        this.productRepository = productRepository;
-    }
-
-    @Transactional
-    public Product create(final ProductRequest productRequest) {
-        return productRepository.save(Product.from(productRequest));
-    }
-    @Transactional(readOnly=true)
-    public List<Product> list() {
-        return productRepository.findAll();
-    }
-
-    @Transactional(readOnly=true)
-    public Product getProduct(Long productId) {
-        return productRepository.findById(productId).orElseThrow(() -> new IllegalArgumentException("상품이 없습니다."));
-    }
-
-}
