--- src/main/java/kitchenpos/product/dto/ProductResponse.java
@@ -1,35 +1,0 @@
-package kitchenpos.product.dto;
-
-import kitchenpos.product.domain.Product;
-
-import java.math.BigDecimal;
-
-public class ProductResponse {
-
-    private final Long id;
-    private final String name;
-    private final BigDecimal price;
-
-    public ProductResponse(Long id, String name, BigDecimal price) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-    }
-
-    public static ProductResponse from(Product product) {
-        return new ProductResponse(product.getId(), product.getName(), product.getPrice());
-    }
-
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-}
