--- src/test/java/kitchenpos/application/unit/table/TableServiceTest.java
@@ -1,110 +1,0 @@
-package kitchenpos.application.unit.table;
-
-import kitchenpos.table.application.TableService;
-import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.dto.OrderTableRequest;
-import kitchenpos.table.repository.OrderTableRepository;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.context.ApplicationEventPublisher;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.*;
-
-@ExtendWith(MockitoExtension.class)
-class TableServiceTest {
-
-    @Mock
-    OrderTableRepository orderTableRepository;
-
-    @Mock
-    ApplicationEventPublisher eventPublisher;
-
-    @DisplayName("테이블을 생성한다.")
-    @Test
-    void createTest() {
-
-        // given
-        OrderTable orderTable = new OrderTable(4, false);
-
-        OrderTable expectedOrderTable = mock(OrderTable.class);
-        when(expectedOrderTable.getId()).thenReturn(1L);
-        when(orderTableRepository.save(orderTable)).thenReturn(expectedOrderTable);
-
-        TableService tableService = new TableService(orderTableRepository, eventPublisher);
-
-        // when
-        OrderTable createdOrderTable = tableService.create(OrderTableRequest.from(orderTable));
-
-        // then
-        assertThat(createdOrderTable.getId()).isNotNull();
-    }
-
-
-    @DisplayName("테이블을 조회한다.")
-    @Test
-    void getListTest() {
-
-        // given
-        OrderTable orderTable = new OrderTable();
-        when(orderTableRepository.findAll()).thenReturn(Arrays.asList(orderTable));
-        TableService tableService = new TableService(orderTableRepository, eventPublisher);
-
-        // when
-        List<OrderTable> orderTables = tableService.list();
-
-        // then
-        assertThat(orderTables).contains(orderTable);
-    }
-
-    @DisplayName("테이블을 비운다.")
-    @Test
-    void changeEmptyTest() {
-
-        //given
-        Long orderTableId = 1L;
-
-        OrderTable orderTable  = mock(OrderTable.class);
-        when(orderTable.isEmpty()).thenReturn(true);
-
-        when(orderTableRepository.findById(orderTableId)).thenReturn(Optional.of(orderTable ));
-
-        OrderTable savedOrderTable = mock(OrderTable.class);
-        when(savedOrderTable.isEmpty()).thenReturn(true);
-
-        TableService tableService = new TableService(orderTableRepository, eventPublisher);
-        // when
-        OrderTable changedOrderTable = tableService.changeEmpty(orderTableId, OrderTableRequest.from(savedOrderTable));
-
-        // then
-        assertThat(changedOrderTable.isEmpty()).isTrue();
-    }
-
-    @DisplayName("테이블의 손님 수를 변경한다.")
-    @Test
-    void changeNumberOfGuests(){
-        // given
-        Long orderTableId = 1L;
-        OrderTable orderTable = mock(OrderTable.class);
-        when(orderTable.getNumberOfGuests()).thenReturn(3);
-        when(orderTableRepository.findById(orderTableId)).thenReturn(Optional.of(orderTable));
-
-        OrderTable savedOrderTable = mock(OrderTable.class);
-        when(savedOrderTable.getNumberOfGuests()).thenReturn(3);
-        when(orderTableRepository.save(orderTable)).thenReturn(savedOrderTable);
-        TableService tableService = new TableService(orderTableRepository, eventPublisher);
-
-        // when
-        OrderTable changedOrderTable = tableService.changeNumberOfGuests(orderTableId, OrderTableRequest.from(orderTable));
-
-        // then
-        assertThat(changedOrderTable.getNumberOfGuests()).isEqualTo(orderTable.getNumberOfGuests());
-    }
-}
